+++ packages/core/src/state-management/__tests__/state-manager.test.ts
@@ -3,7 +3,6 @@ import type { Dispatcher } from '../../dispatcher.js'
 import type { AnchorRequestStore } from '../../store/anchor-request-store.js'
 import type { ExecutionQueue } from '../execution-queue.js'
 import type { DiagnosticsLogger } from '@ceramicnetwork/common'
-import type { ConflictResolution } from '../../conflict-resolution.js'
 import type { LocalIndexApi } from '@ceramicnetwork/indexing'
 import type { RepositoryInternals } from '../repository-internals.js'
 import type { AnchorService } from '../../anchor/anchor-service.js'
@@ -18,7 +17,6 @@ describe('applyWriteOpts', () => {
   const anchorRequestStore = {} as unknown as AnchorRequestStore
   const executionQueue = {} as unknown as ExecutionQueue
   const anchorService = {} as unknown as AnchorService
-  const conflictResolution = {} as unknown as ConflictResolution
   const logger = {} as unknown as DiagnosticsLogger
   const localIndexApi = {} as unknown as LocalIndexApi
   const publishTipFn = jest.fn()
@@ -31,7 +29,6 @@ describe('applyWriteOpts', () => {
     anchorRequestStore,
     executionQueue,
     anchorService,
-    conflictResolution,
     logger,
     localIndexApi,
     internals,