+++ packages/indexing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.2.0...@ceramicnetwork/indexing@1.3.0-rc.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/indexing
+
+
+
+
+
+# [1.2.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.2.0-rc.0...@ceramicnetwork/indexing@1.2.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/indexing
+
+
+
+
+
 # [1.2.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.1.0...@ceramicnetwork/indexing@1.2.0-rc.0) (2023-09-20)
 
 **Note:** Version bump only for package @ceramicnetwork/indexing